# Graph operations performance improvements