--- a/README.md
+++ b/README.md
@@ -1,6 +1,6 @@
 # fp-ts-graph
 
-Immutable, functional graph data structure for [fp-ts](https://github.com/gcanti/fp-ts).
+Immutable, functional, highly performant graph data structure for [fp-ts](https://github.com/gcanti/fp-ts).
 
 ```ts
 type Graph<Id, Edge, Node> = ...
@@ -71,7 +71,7 @@ export const MyIdCodec: Codec<string, string, MyId> = {
     pipe(
       Number(i),
       E.fromPredicate(
-        (parsed: number) => !isNaN(parsed),
+        (parsed: number) => !isNaN(parsed) && i.length > 0,
         () => Dec.error(i, `${i} is not a number`)
       )
     ),