--- a/examples/build-graph.ts
+++ b/examples/build-graph.ts
@@ -1,17 +1,16 @@
-import Graph, * as G from '../src';
-import * as N from 'fp-ts/number';
+import * as G from '../src';
 import { pipe } from 'fp-ts/function';
 import * as O from 'fp-ts/Option';
 import { Option } from 'fp-ts/Option';
 
 // We import our types from the previous section
-import { MyEdge, MyId, MyNode, MyGraph } from './types';
+import { MyEdge, MyId, MyNode, MyGraph, MyIdCodec } from './types';
 
 // To save some writing, we define partially applied versions of the builder functions
 
 const empty = G.empty<MyId, MyEdge, MyNode>();
-const insertNode = G.insertNode(N.Eq);
-const insertEdge = G.insertEdge(N.Eq);
+const insertNode = G.insertNode(MyIdCodec);
+const insertEdge = G.insertEdge(MyIdCodec);
 
 // Then, let's fill the graph with Data.