--- a/src/ImmutableMap.ts
+++ b/src/ImmutableMap.ts
@@ -1,7 +1,8 @@
 /** @since 0.3.0 */
 
+import { Option } from 'fp-ts/Option';
 import * as O from 'fp-ts/Option';
-import * as Enc from 'io-ts/Encoder';
+import { Encoder } from 'io-ts/Encoder';
 import { Map } from 'immutable';
 import { pipe } from 'fp-ts/function';
 
@@ -12,7 +13,7 @@ import { pipe } from 'fp-ts/function';
  * @category Combinators
  */
 export const upsertAt =
-  <K>(E: Enc.Encoder<string, K>) =>
+  <K>(E: Encoder<string, K>) =>
   <A>(k: K, a: A) =>
   (m: Map<string, A>): Map<string, A> => {
     const encodedKey = E.encode(k);
@@ -29,9 +30,9 @@ export const upsertAt =
  * @category Combinators
  */
 export const modifyAt =
-  <K>(E: Enc.Encoder<string, K>) =>
+  <K>(E: Encoder<string, K>) =>
   <A>(k: K, f: (a: A) => A) =>
-  (m: Map<string, A>): O.Option<Map<string, A>> => {
+  (m: Map<string, A>): Option<Map<string, A>> => {
     const encodedKey = E.encode(k);
     return pipe(
       m.get(encodedKey, null),
@@ -47,7 +48,7 @@ export const modifyAt =
  * @category Utils
  */
 export const lookup =
-  <K>(E: Enc.Encoder<string, K>) =>
+  <K>(E: Encoder<string, K>) =>
   (k: K) =>
-  <A>(m: Map<string, A>): O.Option<A> =>
+  <A>(m: Map<string, A>): Option<A> =>
     pipe(m.get(E.encode(k), null), O.fromNullable);