Improves overall library performance by changing underlying data structures to [immutable.js Map](https://immutable-js.com/) with arbitrary keys encoded to string using [io-ts](https://gcanti.github.io/io-ts/) [Codec](https://gcanti.github.io/io-ts/modules/Codec.ts.html). See #14 for reasoning behind the change.

Before / after benchmark for 10k operations
```
                        0.2.2 (Map)       0.3.0-dev (immutable.js Map + io-ts)
Insert 10000 nodes        6517 ms               59 ms
Insert 10000 edges       39047 ms               85 ms
Lookup 10000 nodes        1985 ms                8 ms
Lookup 10000 edges        4848 ms               14 ms
```